Move table to diff-diveinfo-results component
