+++ projects/planner/src/app/diff/diveinfo/results/diff-diveinfo-results.component.spec.ts
@@ -1,6 +1,11 @@
 import {ComponentFixture, TestBed} from '@angular/core/testing';
 
 import {DiveInfoResultsDifferenceComponent} from './diff-diveinfo-results.component';
+import {ViewSwitchService} from '../../../shared/viewSwitchService';
+import {DiveSchedules} from '../../../shared/dive.schedules';
+import {ReloadDispatcher} from '../../../shared/reloadDispatcher';
+import {UnitConversion} from '../../../shared/UnitConversion';
+import {ProfileComparatorService} from '../../../shared/profileComparatorService';
 
 describe('DiveInfoResultsDifferenceComponent', () => {
     let component: DiveInfoResultsDifferenceComponent;
@@ -8,7 +13,14 @@ describe('DiveInfoResultsDifferenceComponent', () => {
 
     beforeEach(() => {
         TestBed.configureTestingModule({
-            declarations: [DiveInfoResultsDifferenceComponent]
+            declarations: [DiveInfoResultsDifferenceComponent],
+            providers: [
+                ViewSwitchService,
+                DiveSchedules,
+                ReloadDispatcher,
+                UnitConversion,
+                ProfileComparatorService
+            ]
         });
         fixture = TestBed.createComponent(DiveInfoResultsDifferenceComponent);
         component = fixture.componentInstance;
