+++ projects/planner/src/app/diff/diveinfo/results/diff-diveinfo-results.component.ts
@@ -1,4 +1,9 @@
 import {Component} from '@angular/core';
+import {ViewSwitchService} from '../../../shared/viewSwitchService';
+import {UnitConversion} from '../../../shared/UnitConversion';
+import {ProfileComparatorService} from '../../../shared/profileComparatorService';
+import {DiveResults} from '../../../shared/diveresults';
+import {formatNumber} from '@angular/common';
 
 @Component({
     selector: 'app-diff-diveinfo-results',
@@ -7,4 +12,47 @@ import {Component} from '@angular/core';
 })
 export class DiveInfoResultsDifferenceComponent {
 
+    constructor(
+        private viewSwitch: ViewSwitchService,
+        public units: UnitConversion,
+        private profileComparatorService: ProfileComparatorService) {
+    }
+    public get dive(): DiveResults {
+        return this.profileComparatorService.profileAResults();
+    }
+
+    public get isComplex(): boolean {
+        return this.viewSwitch.isComplex;
+    }
+
+    public get showMaxBottomTime(): boolean {
+        return this.dive.maxTime > 0;
+    }
+
+    public get noDeco(): number {
+        return this.dive.noDecoTime;
+    }
+
+    public get averageDepth(): number {
+        return this.units.fromMeters(this.dive.averageDepth);
+    }
+
+    public get highestDensity(): number {
+        const density = this.dive.highestDensity.density;
+        return this.units.fromGramPerLiter(density);
+    }
+
+    public get densityText(): string {
+        const gas = this.dive.highestDensity.gas.name;
+        const depth = this.units.fromMeters(this.dive.highestDensity.depth);
+        return `${gas} at ${depth} ${this.units.length}`;
+    }
+
+    public get cnsText(): string {
+        if(this.dive.cns >= 1000) {
+            return '> 1000';
+        }
+
+        return formatNumber(this.dive.cns, 'en', '1.0-0');
+    }
 }
